Most tests passing, now tackle the big refactor
